Almost done
Reorganized archives.
On home created a search component. Used useEffect to create item filter for description and color.
Fetched all the products from the given address and made a separated component for it.
Also created separated archives for checkout page and success transaction page. Made a separated component for payment methods in checkout page and for the progress bar on top of the page.